+++ libcloud/compute/drivers/openstack.py
@@ -3908,7 +3908,7 @@ class OpenStack_2_NodeDriver(OpenStack_1_1_NodeDriver):
 
         return res
 
-    def get_quota_set(self, tenant_id, user_id=None):
+    def ex_get_quota_set(self, tenant_id, user_id=None):
         """
         Get the quota for a project or a project and a user.
 
@@ -4219,10 +4219,22 @@ class OpenStack_2_PortInterface(UuidMixin):
 
 class OpenStack_2_QuotaSetItem(object):
     """
-    A Qouta Set Item.
+    Qouta Set Item info. Each item has three attributes: in_use,
+    limit and reserved.
+
+    See:
+    https://docs.openstack.org/api-ref/compute/?expanded=show-the-detail-of-quota-detail#show-a-quota
     """
 
     def __init__(self, in_use, limit, reserved):
+        """
+        :param in_use: Number of currently used resources.
+        :type in_use: ``int``
+        :param limit: Max number of available resources.
+        :type limit: ``int``
+        :param reserved: Number of reserved resources.
+        :type reserved: ``int``
+        """
         self.in_use = in_use
         self.limit = limit
         self.reserved = reserved
@@ -4235,7 +4247,11 @@ class OpenStack_2_QuotaSetItem(object):
 
 class OpenStack_2_QuotaSet(object):
     """
-    A Qouta Set.
+    Qouta Set info. To get the informatio about quotas and used resources.
+
+    See:
+    https://docs.openstack.org/api-ref/compute/?expanded=show-the-detail-of-quota-detail#show-a-quota
+
     """
 
     def __init__(self, id, cores, instances, key_pairs, metadata_items, ram,
@@ -4244,6 +4260,44 @@ class OpenStack_2_QuotaSet(object):
                  security_groups=None, injected_file_content_bytes=None,
                  injected_file_path_bytes=None, injected_files=None,
                  driver=None):
+        """
+        :param id: Quota Set ID.
+        :type id: ``str``
+        :param cores: Quota Set of cores.
+        :type cores: :class:`.OpenStack_2_QuotaSetItem` or ``dict``
+        :param instances: Quota Set of instances.
+        :type instances: :class:`.OpenStack_2_QuotaSetItem` or ``dict``
+        :param key_pairs: Quota Set of key pairs.
+        :type key_pairs: :class:`.OpenStack_2_QuotaSetItem` or ``dict``
+        :param metadata_items: Quota Set of metadata items.
+        :type metadata_items: :class:`.OpenStack_2_QuotaSetItem` or ``dict``
+        :param ram: Quota Set of RAM.
+        :type ram: :class:`.OpenStack_2_QuotaSetItem` or ``dict``
+        :param server_groups: Quota Set of server groups.
+        :type server_groups: :class:`.OpenStack_2_QuotaSetItem` or ``dict``
+        :param fixed_ips: Quota Set of fixed ips. (optional)
+        :type fixed_ips: :class:`.OpenStack_2_QuotaSetItem` or ``dict``
+        :param floating_ips: Quota Set of floating ips. (optional)
+        :type floating_ips: :class:`.OpenStack_2_QuotaSetItem` or ``dict``
+        :param networks: Quota Set of networks. (optional)
+        :type networks: :class:`.OpenStack_2_QuotaSetItem` or ``dict``
+        :param security_group_rules: Quota Set of security group rules.
+                                     (optional)
+        :type security_group_rules: :class:`.OpenStack_2_QuotaSetItem`
+                                    or ``dict``
+        :param security_groups: Quota Set of security groups. (optional)
+        :type security_groups: :class:`.OpenStack_2_QuotaSetItem` or ``dict``
+        :param injected_file_content_bytes: Quota Set of injected file content
+                                            bytes. (optional)
+        :type injected_file_content_bytes: :class:`.OpenStack_2_QuotaSetItem`
+                                           or ``dict``
+        :param injected_file_path_bytes: Quota Set of injected file path bytes.
+                                         (optional)
+        :type injected_file_path_bytes: :class:`.OpenStack_2_QuotaSetItem`
+                                        or ``dict``
+        :param injected_files: Quota Set of injected files. (optional)
+        :type injected_files: :class:`.OpenStack_2_QuotaSetItem` or ``dict``
+        """
         self.id = str(id)
         self.cores = self._to_quota_set_item(cores)
         self.instances = self._to_quota_set_item(instances)
@@ -4268,8 +4322,11 @@ class OpenStack_2_QuotaSet(object):
 
     def _to_quota_set_item(self, obj):
         if obj:
-            return OpenStack_2_QuotaSetItem(obj['in_use'], obj['limit'],
-                                            obj['reserved'])
+            if isinstance(obj, OpenStack_2_QuotaSetItem):
+                return obj
+            elif isinstance(obj, dict):
+                return OpenStack_2_QuotaSetItem(obj['in_use'], obj['limit'],
+                                                obj['reserved'])
         else:
             return None